Factor out filter logic to a util
